feat: refactor image rendering and move terminal image functionality to a new module
